requirements.txt
django-cors-headers==4.3.1
python-dotenv==1.0.0